refactor: move network infrastructure under core data
