+++ lib/core/README.md
@@ -12,9 +12,9 @@ Here is a breakdown of the typical subdirectories found in `core/`:
 - **`config/`**: App-wide configurations, environment variables, feature flags, and constants.
 - **`di/`**: Dependency Injection setup (e.g., `get_it` or `injectable` configurations) for registering all core services, repositories, and BLoCs.
 - **`domain/`**: Architecture-level domain primitives shared by feature domains, such as `Failure`, `Result<T>`, and `UseCase`.
+- **`data/`**: Reusable data-layer infrastructure, such as networking clients, interceptors, API configuration, and data-source helpers.
 - **`extensions/`**: Generic extensions used across the app (e.g., `build_context_extensions.dart`, `future_extensions.dart`).
 - **`layout/`**: Responsive layout primitives such as breakpoints.
-- **`network/`**: Base networking clients (e.g., `Dio` or `http`), interceptors, API configuration, and network connectivity checkers.
 - **`platform/`**: App-wide platform and plugin integrations, such as Firebase initialization, media picking/playback, push/local notifications, OS permissions, and sharing.
 
 > [!NOTE]
